move uname_result
