+++ stdlib/os/__init__.pyi
@@ -396,7 +396,6 @@ def strerror(__code: int) -> str: ...
 def umask(__mask: int) -> int: ...
 
 if sys.platform != "win32":
-    # Unix only
     def ctermid() -> str: ...
     def getegid() -> int: ...
     def geteuid() -> int: ...
@@ -426,7 +425,13 @@ if sys.platform != "win32":
     def getsid(__pid: int) -> int: ...
     def setsid() -> None: ...
     def setuid(__uid: int) -> None: ...
-    from posix import uname_result
+    @final
+    class uname_result(NamedTuple):
+        sysname: str
+        nodename: str
+        release: str
+        version: str
+        machine: str
     def uname() -> uname_result: ...
 
 @overload
